docs/src/requirements.txt
azure-ai-documentintelligence
azure.core
streamlit
azure-storage-blob
python-dotenv